+++ lib/qiniu/utils.rb
@@ -74,10 +74,11 @@ module Qiniu
         else
           data = {}
           body = response.respond_to?(:body) ? response.body : {}
+          raw_headers = response.respond_to?(:raw_headers) ? response.raw_headers : {}
           data = safe_json_parse(body) unless body.empty?
         end
-        [code, data]
-      end
+        [code, data, raw_headers]
+      end # send_request_with
 
       def http_request url, data = nil, options = {}
         retry_times = 0
@@ -99,9 +100,10 @@ module Qiniu
             res = e.response
             code = res.code.to_i if res.respond_to? :code
             body = res.respond_to?(:body) ? res.body : ""
+            raw_headers = res.respond_to?(:raw_headers) ? res.raw_headers : {}
             data = safe_json_parse(body) unless body.empty?
           end
-          [code, data]
+          [code, data, raw_headers]
         end
       end