+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## CHANGE LOG
 
+### v6.1.0
+
+- Qiniu::Storage所有上传接口返回第三个值raw_headers，类型为Hash，包含已解析的HTTP响应报文中的所有Header信息。
+
+该返回值主要用于调试。当遇到难以理解或解释的错误时，请将其中的X-Log和X-Reqid两项信息[通过邮件反馈](mailto:support@qiniu.com?subject=Ruby-SDK-Bug-Report)给我们。
+
+- 更新Qiniu::Storage所有上传接口的测试用例，打印HTTP响应Header信息。
+
+- 删除过时的Qiniu::Storage#put_file方法和相关测试用例。
+
+该方法调用的API已过时并逐步废弃，建议用户尽快迁移到Qiniu::Storage#upload_with_token_2方法上。
+
 ### v6.0.1
 
 - 重新划分命名空间，存储相关归入Qiniu::Storage，数据处理相关归入Qiniu::Fop，杂项相关归入Qiniu::Misc。